+++ examples/chapi-example/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @learncard/chapi-example
 
+## 1.0.42
+
+### Patch Changes
+
+-   Updated dependencies [[`d29a2d5`](https://github.com/learningeconomy/LearnCard/commit/d29a2d5dd0d4ac1c7f178dc0d00956561012a1fb), [`2707350`](https://github.com/learningeconomy/LearnCard/commit/27073505a7e0c4f4ec8fef191fbff16d97c5d74a), [`c997138`](https://github.com/learningeconomy/LearnCard/commit/c9971386188169859bbe9e8ef7da6572432ed3f0), [`8fbee11`](https://github.com/learningeconomy/LearnCard/commit/8fbee11a07df86ba7d3fdee7430a7288b6ba321f)]:
+    -   @learncard/react@2.5.1
+
 ## 1.0.41
 
 ### Patch Changes